Refactor Pokemon object (Pokemon extends PokemonDetails), moved a bunch of stuff to PokemonDetails, made NormalEncounterResult and DiskEncounterResult extend PokemonDetails. Modifed the map cache timer to 10 seconds, fixed the pokemon catchable cache not refreshing.
